+++ Changelog.md
@@ -33,6 +33,9 @@
   ([#344](https://github.com/tafia/quick-xml/issues/344))
 - test: add tests for trivial documents (empty / only comment / `<root>...</root>` -- one tag with content)
 - fix: CDATA was not handled in many cases where it should
+- fix: do not unescape CDATA content because it never escaped by design.
+  CDATA event data now represented by its own `BytesCData` type
+  ([quick-xml#311](https://github.com/tafia/quick-xml/issues/311))
 
 ## 0.23.0-alpha3